Add a generic helper method to create new proxy instances.
This change isolates all knowledge of the Proxy and InvocationHandler
interfaces to the Proxies helper class. The tests just create a proxy
that implements the desired interface and uses reflection to delegate
to the given target object:

    Proxies.newProxyInstance(AdaptorContext.class, new AdaptorContextMock());

    private class AdaptorContextMock {
      public Config getConfig() {
        return new Config();
      }
    }

This was a previously rejected idea, but I wanted to show it action,
and how clean and readable the tests were as a result.

Code review: https://codereview.appspot.com/198430043/