Properly encode the Link URLs in Folder Listings
This change makes sure the URLs supplied in Folder listings
are properly URL-encoded. This is done using the same
mechanisms as the filesytem adaptor and opentext adaptor,
including copying the HtmlResponseWriter implementation
from the filesystem adaptor and removing the partial
implementation that was embedded in the DocumentumAdaptor
class.

Code Review: http://codereview.appspot.com/223810043