Adds H2 Embedded Database for Use in Tests
This change adds the H2 embedded database to the project for use
in the tests.  Certain functions of the adaptor require running
DQL queries against the Documentum database.  The tests can mock
the Documentum database tables in H2, and inject them with test
data.

This is done as a separate change because the ACL tests and the
Groups tests both need it and they are being written concurrently.

This also imports the JdbcFixture class from the Documentum v3
Connector with minor refactoring to simplify connection management
and extract the Documentum-specific DDL. This will allow the class
to be more easily used in other adaptors.

TODO: Refactor some of the existing SysObj tests to use H2 as
the backing store, rather than Maps in the Proxies.

Code Review: http://codereview.appspot.com/223640043