Code changes to support Documentum TCS required groups and required group sets in ACLs. Code checks for permit_type to determine if the group is required group or required group set.
For required group, all the groups are required. An intermediary ACL is created
for each required group with and-both-permit inheritance type. A final ACL with
actual principals is created with inherit-from the intermediary ACL.

For required group set, one of the groups in the group set is required group.
An intermediary ACL is created with all the groups of the group set with
and-both-permit inheritance type. A final ACL with actual principals is
created with inherit-from the intermediary ACL.

Code review: https://codereview.appspot.com/230490043/