Refactor Tests: Use Set.equals(), eliminates redundant size checks, etc.
This refactoring eliminates redundant assertions in the tests.
These redundancies typically looked like:

  Acl aclList = namedResources.get(new DocId("4501081f80000100"));
  Set<UserPrincipal> allowUsers = aclList.getPermitUsers();
  Set<UserPrincipal> denyUsers = aclList.getDenyUsers();

  assertEquals(2, allowUsers.size());
  assertEquals(1, denyUsers.size());

  UserPrincipal user2 = new UserPrincipal("User2");
  assertFalse(allowUsers.contains(user2));

  assertEquals(ImmutableSet.of(new UserPrincipal("User4", "globalNS"),
      new UserPrincipal("User5", "globalNS")), allowUsers);
  assertEquals(ImmutableSet.of(new UserPrincipal("User2", "globalNS")),
      denyUsers);

In this example, the checks for the size of the set are
useless, because the check for set equality will fail if
the sets are different sizes.

The check for User2 not being in the results would also
be caught as it is not in the expected set. Even if there
was a logic failure in the ACL code that did include User2,
the check here will always pass, because it should have
constructed the expected UserPrincipal as:
  new UserPrincipal("User2", "globalNS");

The improved test code lookes like:

  Acl acl = namedResources.get(new DocId("4501081f80000100"));
  assertEquals(ImmutableSet.of(new UserPrincipal("User4", "globalNS"),
      new UserPrincipal("User5", "globalNS")), acl.getPermitUsers());
  assertEquals(ImmutableSet.of(new UserPrincipal("User2", "globalNS")),
      acl.getDenyUsers());
  assertEquals(ImmutableSet.of(), acl.getPermitGroups());
  assertEquals(ImmutableSet.of(), acl.getDenyGroups());

This code eliminates the redundant size checks, the redundant
not contains check, the temporary variables that were used
for those checks, and relies on Set.equals() to do the check.
It also improves the test, because it verifies that no
permitted or denied groups were added (although some tests
did check this by looking for a size() of 0).

Finally, I modified a bunch of tests to declare they throw
Exception, rather than a list of specific Exceptions.
This was done to reduce line wrapping.

Code Review: http://codereview.appspot.com/237210043