Refactors pushAclUpdates to take clientX, session, principals, and checkpoint
This change modifies DocumentumAdaptor.pushAclUpdates() to take the
clientX, session, principals, and checkpoint as parameters, and return
a new checkpoint.  This more closely follows the pushdDocumentUpdates()
implementation and makes getModidfiedDocIds() look more like getDocIds().

This change replaces the internally stored aclModifiedDate and aclModifyId
fields in DocumentumAcls with a Checkpoint that is passed in. A new
getAclUpdatesCheckpoint() method was added to retrieve the Checkpoint that
was generated by a previous call to getUpdateAcls().

Finally, the alcUpdate tests were updated to supply and validate
returned checkpoints.

Code Review: http://codereview.appspot.com/239480043