Refactor Tests: Make ACLMock a true fake against H2 DB
ACLMock was being used as a mechanism to populate the database
with ACLs during test setup, in addition to being returned from
ACL lookups from the production code.

This change breaks the functionality of creating the ACLs
in the database out into private methods called directly
by the tests during test environment setup.  These methods now
take the ID of the ACL rather than the ACLMock object itself,
so a bit of clean-up was done with regard to the ID string
constants.

Finally, now that ACLMocks were no longer used in test setup,
the instantiation of the AclTestProxies could be pushed down
into getAllAcls().

Code Review: http://codereview.appspot.com/241290043