Support / as a startPath to list all cabinets.
This change adds support for specifying / as a startPath,
which we expand to all Cabinets in the docbase during a crawl
request of "/".

This effectively feeds all documents and folders in the docbase.

This change imports the MockRequest and MockResponse objects from
the Filesystem Adaptor so that it may use the H2BackedTestProxies,
rather than yet another variation of the GetDocContentTestProxies.

Code Review: http://codereview.appspot.com/243510043