RespondNotModified for objects not modified since last crawl.
Return a response of not modified if the Documentum object's
last modified time is not sufficiently recent with regards to
the request's last access time.

Dates from Documentum can be in the server's local time,
so only count a object as unmodified if its last modified
time is more than a day before the last crawl time.

Code Review:  http://codereview.appspot.com/244560043