Refactor Tests: getDocContent tests call adaptor.init(), adaptor.getDocContent()
This change converts most of the getDocContent tests to call the public
adaptor.getDocContent() method, rather than getDocContentHelper() with
its ever-expanding list of parameters.  To accomplish this, the
H2BackedTest proxies were enhanced to allow adaptor.init() to be called.
This allows the tests to set the appropriate configuration parameters in the
Config properties, call adaptor.init(), then call adaptor.getDocContent().

The lone straggler for now is checkGetRootContent() which needs to pass
in a variable maxHtmlLinks parameter.

Code Review: http://codereview.appspot.com/255810043