Refactor Tests: Tests call adaptor.getDocIds() rather than helper methods
This change forces the Groups and ACLs tests to call
adaptor.getDocIds(), rather than the various helper methods it called.
The tests now set appropriate configuration paramenters, call
adaptor.init(), then adaptor.getDocIds(), and finally extract the
collected output from the recording pusher.

The adaptor derives the local namespace from the adaptor.namespace
and the docbase name. The tests now set the adaptor.namespace to
"NS", and the docbase name to "Local".  This results in a global
namespace of "NS" (instead of "globalNS"), and a local namespace
of "NS_Local" (instead of "localNS").

This change also adds tests for getDocIds() pushing the start paths,
which was never tested in the past.

Code Review: http://codereview.appspot.com/254060043