Refactor retries in getDocIds for code sharing and testability.
Add a template method for traversing ACLs and groups. Each subclass
must implement methods to create a collection of objects, to fill the
collection from the repository, and to push the collection to the GSA.
Each instance is long-lived and encapsulates a checkpoint.

Add a test subclass to replay checkpoint and exception scenarios.

Related changes:

* Declare previousCheckpoint and caughtException inside the loop for
  simplicity. (This also guarantees the setting of previousCheckpoint
  if we add batching.)

* Use a Deque to create a chain of suppressed exceptions. (This also
  avoids double logging of exceptions in some circumstances.)

* In getGroups and pushGroupUpdates, initialize groupName to null
  rather than empty string. The empty string was propagated to the
  checkpoint and then in some circumstances to the groups query as
  group_name > ''. Nothing was broken, but it looked odd in the logs.

Other changes:

* Refactor the test methods that get a DocumentumAdaptor instance.
* Fix lint errors.

Code review: https://codereview.appspot.com/310880043/