GroupTraverser supports batching the results.
We could not use both ROW_BASED and RETURN_TOP in our
groups query as the RETURN_TOP counts the expanded
rows. Consequently when batching group lookup,
I run two queries. The first just fetches the
a batch size of object IDs of the groups.
The first and last object IDs in the returned
result are then used to construct a BETWEEN clause
for the actual group lookup with members.

The makeGroupsQuery() method was used by the
GroupsTraverser and the ModifiedGroupsTraverser.
However, once changes were made to support batching,
makeGroupsQuery() grew to be an ugly mess.
Therefore, I split it into makeGroupsQuery() for
full traversals (with batch support),
and makeModifiedGroupsQuery() for incremental
traversals (and no batch support).

Code Review: http://codereview.appspot.com/308590043