Treat a DfIdNotFoundException as a deleted ACL.
Previously, if we got a pair of dm_save, dm_destroy events, we would
skip the DfIdNotFoundException on the dm_save, avoid caching the
event, and then send an empty ACL on the dm_destroy. Now, we send an
empty ACL on the DfIdNotFoundException, cache the event, and skip the
dm_destroy as redundant.

Related changes:

* Remove event_name from the ORDER BY clause. This was a carryover
  from v3, and it was always broken. It was meant to process deletes
  before saves, but that isn't possible, since saves happen earlier.

* Reduce the log level of the action taken on each ACL audit event.

Other changes:

* Add checks for different permissions to testUpdateAclsPrincipals.

Code review: https://codereview.appspot.com/312860043/