Add Traverser Exception tests for modified ACLs.
These replace the hacky tests relying on the AssertionError in
SessionMock.getObject.

Extend ExceptionalResultSetTestProxies to take an iterator of fail
iterations. This allows full control over when exceptions are thrown.
The existing model of a fixed fail iteration is supported with a
default cyclic iterator over a single given integer.

Code review: https://codereview.appspot.com/314830043/