Fix the Content-Type header (b/32103403)
The adaptor was returning the documentum-specific
dm_format name as the response Content-Type header.
This change calls IDfFormat.getMIMEType() to
fetch the actual MIME type for the content.

This change also matches the v3 connector behaviour
by not setting the contentType in the response if
no actual content is being returned.

Code Review: http://codereview.appspot.com/314870043