Fix getDocContent for non-root start paths
The fix for bug#26742691 broke getDocContent() for non-root start paths.
The start path with an object ID suffix failed the isUnderStartPath() test.
This modifies isUnderStartPath() to take a DocId instead of a path to ensure
we correctly consider only the path portion of the DocId.

Code Review: https://codereview.appspot.com/317120043/