Fix b/32392655 Handle documents with a slash in their names
Documentum allows documents (but not folders) to contain a slash
character in their names. The GSA interprets the slash as a path
separator and screws up index diagnostics. Similarly, Documentum's
getObjectByPath() fails to find the document. Since the adaptor now
includes the objectid to the docid and uses that to look up the document,
the path part is no longer as important. This change URL encodes slashes
as %2F in object names, so the subsequent DocIds will not contain slashes
that are not path separators.

Code Review: http://codereview.appspot.com/311550043/