Remove redundant setParentFolderId. (#3)
Part of a series of changes removing redundant inserts and updates.

The change in testModifiedDocumentsNoCheckpointObjId to expect the
folder as well as the aaa and bbb documents restores the original
behavior, which was modified in commit dea56c3a. Apparently, the
length of the object ID has some impact on the query results from H2.